Don't include <stdint.h> and use types from <stdint.h>. These types are not portable w/r format strings. Instead we standardise on using (unsigned) long long instead of uint64_t and int64_t respectively etc. Also remove any usage of <inttypes.h> and its horrible macros.

+++ libmonit/src/io/OutputStream.c
@@ -64,7 +64,7 @@ struct T {
         uchar_t *length;
         bool isclosed;
         int sessionWritten;
-        int64_t bytesWritten;
+        long long bytesWritten;
         uchar_t buffer[BUFFER_SIZE + 1];
 };
 
@@ -392,7 +392,7 @@ bool OutputStream_isClosed(T S) {
 }
 
 
-int64_t OutputStream_getBytesWritten(T S) {
+long long OutputStream_getBytesWritten(T S) {
         assert(S);
         return S->bytesWritten;
 }
