+++ libmonit/src/system/System.c
@@ -130,8 +130,8 @@ bool System_random(void *buf, size_t nbytes) {
 }
 
 
-uint64_t System_randomNumber() {
-        uint64_t random;
+unsigned long long System_randomNumber() {
+        unsigned long long random;
         System_random(&random, sizeof(random));
         return random;
 }
