+++ libmonit/src/system/Time.h
@@ -119,7 +119,7 @@ time_t Time_now(void);
  * Coordinated Universal Time (UTC).
  * @exception AssertException If time could not be obtained
  */
-int64_t Time_milli(void);
+long long Time_milli(void);
 
 
 /**
@@ -129,7 +129,7 @@ int64_t Time_milli(void);
  * Coordinated Universal Time (UTC).
  * @exception AssertException If time could not be obtained
  */
-int64_t Time_micro(void);
+long long Time_micro(void);
 
 
 /**
